Only load the necessary page or layout per call
Which means we don't do any extra work apart from reading the
filesystem right now. So if we were more efficient with the filesystem
we could make this better.

Although, doing it more efficient for the filesystem is not needed for
a cached prod version, but avoiding all those calls for a dev version
would be nice…